weblibrary/ai-service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
google-genai==1.16.1
# faiss-cpu==1.11.0
# sentence-transformers==4.1.0